@@ -3,10 +3,10 @@ description: "Deliver for Delivering Services"
 
 publish_to: "none"
 
-version: 3.1.0+31
+version: 3.1.0+32
 
 environment:
-  sdk: ^3.7.2
+  sdk: ^3.9.0
 
 dependencies:
   flutter:
@@ -21,43 +21,44 @@ dependencies:
   bloc_concurrency: ^0.3.0
   equatable: ^2.0.7
   dartz: ^0.10.1
-  dio: ^5.8.0+1
-  connectivity_plus: ^6.1.4
-  get_it: ^8.0.3
-  injectable: ^2.5.0
-  built_value: ^8.10.1
+  dio: ^5.9.0
+  connectivity_plus: ^6.1.5
+  get_it: ^8.2.0
+  injectable: ^2.5.1
+  built_value: ^8.11.1
   built_collection: ^5.1.1
-  flutter_svg: ^2.1.0
+  flutter_svg: ^2.2.1
   fluttertoast: ^8.2.12
-  flutter_spinkit: ^5.2.1
+  flutter_spinkit: ^5.2.2
   json_annotation: ^4.9.0
   shared_preferences: ^2.5.3
   flutter_secure_storage: ^9.2.4
-  firebase_core: ^3.14.0
-  firebase_storage: ^12.4.7
-  firebase_messaging: ^15.2.5
-  flutter_local_notifications: ^19.2.0
-  location: ^8.0.0
-  google_maps_flutter: ^2.12.1
-  pinput: ^5.0.1
+  firebase_core: ^4.1.0
+  firebase_storage: ^13.0.1
+  firebase_messaging: ^16.0.1
+  cloud_firestore: ^6.0.1
+  flutter_local_notifications: ^19.4.1
+  location: ^8.0.1
+  google_maps_flutter: ^2.13.1
+  pinput: ^5.0.2
   sliding_up_panel: ^2.0.0+1
-  geocoding: ^3.0.0
-  flutter_dotenv: ^5.2.1
+  geocoding: ^4.0.0
+  flutter_dotenv: ^6.0.0
   auto_size_text: ^3.0.0
-  image_picker: ^1.1.2
+  image_picker: ^1.2.0
   shimmer: ^3.0.0
   fancy_shimmer_image: ^2.0.3
-  device_info_plus: ^11.4.0
+  device_info_plus: ^11.5.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^5.0.0
-  build_runner: ^2.4.15
-  built_value_generator: ^8.9.5
-  injectable_generator: ^2.7.0
-  json_serializable: ^6.9.5
+  flutter_lints: ^6.0.0
+  build_runner: ^2.7.1
+  built_value_generator: ^8.11.1
+  injectable_generator: ^2.8.1
+  json_serializable: ^6.11.1
 
 flutter:
   uses-material-design: true
