@@ -3,8 +3,8 @@ import 'package:injectable/injectable.dart';
 
 import '../../../../core/error/failures.dart';
 import '../../../../core/repositories/base_repository_impl.dart';
-import '../../domain/entities/profile.dart';
 import '../data_sources/profile_remote_data_source.dart';
+import '../models/profile.dart';
 
 @lazySingleton
 class ProfileRepository extends BaseRepositoryImpl {
@@ -13,9 +13,16 @@ class ProfileRepository extends BaseRepositoryImpl {
   ProfileRepository(this._profileRemoteDataSource, super._networkInfo);
 
   Future<Either<Failure, Profile>> getProfile() async => await requestApi(
-        () async => await _profileRemoteDataSource.getProfile(),
-        (profileModel) {
-          return profileModel.toDomain();
-        },
-      );
+    () async => await _profileRemoteDataSource.getProfile(),
+    (profileModel) {
+      return profileModel;
+    },
+  );
+
+  Future<Either<Failure, String?>> logout() async => await requestApi(
+    () async => await _profileRemoteDataSource.logout(),
+    (message) {
+      return message;
+    },
+  );
 }
