@@ -9,10 +9,9 @@ part of 'splash_event.dart';
 class _$CheckUserAuthentication extends CheckUserAuthentication {
   factory _$CheckUserAuthentication([
     void Function(CheckUserAuthenticationBuilder)? updates,
-  ]) => (new CheckUserAuthenticationBuilder()..update(updates))._build();
+  ]) => (CheckUserAuthenticationBuilder()..update(updates))._build();
 
   _$CheckUserAuthentication._() : super._();
-
   @override
   CheckUserAuthentication rebuild(
     void Function(CheckUserAuthenticationBuilder) updates,
@@ -20,7 +19,7 @@ class _$CheckUserAuthentication extends CheckUserAuthentication {
 
   @override
   CheckUserAuthenticationBuilder toBuilder() =>
-      new CheckUserAuthenticationBuilder()..replace(this);
+      CheckUserAuthenticationBuilder()..replace(this);
 
   @override
   bool operator ==(Object other) {
@@ -48,7 +47,6 @@ class CheckUserAuthenticationBuilder
 
   @override
   void replace(CheckUserAuthentication other) {
-    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$CheckUserAuthentication;
   }
 
@@ -61,7 +59,7 @@ class CheckUserAuthenticationBuilder
   CheckUserAuthentication build() => _build();
 
   _$CheckUserAuthentication _build() {
-    final _$result = _$v ?? new _$CheckUserAuthentication._();
+    final _$result = _$v ?? _$CheckUserAuthentication._();
     replace(_$result);
     return _$result;
   }
